+++ src/Cwaff-Guns/Flakseed.cs
@@ -5,7 +5,7 @@ public class Flakseed : CwaffGun
 {
     public static string ItemName         = "Flakseed";
     public static string ShortDescription = "Orgunic Gardening";
-    public static string LongDescription  = "Fires flak seeds that grow into flak flowers 6 seconds after landing. Bullets accelerate growth by pollinating flak sprouts they pass over, with stronger bullets causing faster growth. Grown flowers fire flak for 15 seconds before withering. Nearby flak flowers compete for nutrients and accelerate withering. Flak flowers will not wither while planted in water, and wither instantly when exposed to hostile terrain or when trampled by enemies.";
+    public static string LongDescription  = "Fires flak seeds that grow into flak flowers 3 seconds after landing. Bullets accelerate growth by pollinating flak sprouts they pass over, with stronger bullets causing faster growth. Grown flowers fire flak for 15 seconds before withering. Flak flowers will not wither while planted in water, and wither instantly when exposed to hostile terrain or when trampled by enemies.";
     public static string Lore             = "The quintessential tool for both practitioners of warfare-based gardening and practitioners of gardening-based warfare. Disregarding the fact that the combined demographic for both of these hobbies was 0 at the time of this tool's invention, a successful decade-long marketing campaign has since brought that number up to 2. With just a tiny bit of practice, you could be the one to let them technically, legally claim that number is 3!";
 
     internal static GameObject _FlakFlowerPrefab     = null;
@@ -107,7 +107,7 @@ public class FlakseedProjectile : MonoBehaviour
 
 public class FlakseedFlower : MonoBehaviour
 {
-    const float _GROWTH_TIME = 6.0f;
+    const float _GROWTH_TIME = 3f;
     const float _MIN_FIRE_RATE = 0.5f;
     const float _FIRE_RATE_VARIANCE = 0.25f;
     const float _WILT_TIME = 15.0f;
@@ -215,6 +215,25 @@ public class FlakseedFlower : MonoBehaviour
             startScale: 1.0f,
             endScale: 0.1f,
             randomFrame: true
+          );
+    }
+
+    private void DoTrampleVFX()
+    {
+        base.gameObject.Play("trample_sound");
+        CwaffVFX.SpawnBurst(
+            prefab           : Groundhog._EarthClod,
+            numToSpawn       : 11,
+            basePosition     : this._firePos,
+            positionVariance : 0.25f,
+            velocityVariance : 4f,
+            velType          : CwaffVFX.Vel.AwayRadial,
+            rotType          : CwaffVFX.Rot.Random,
+            lifetime         : 0.35f,
+            fadeOutTime      : 0.1f,
+            // uniform          : true,
+            startScale       : 1.0f,
+            endScale         : 0.2f
           );
     }
 
@@ -343,8 +362,8 @@ public class FlakseedFlower : MonoBehaviour
 
     private float RateOfWilting()
     {
-        const float MAX_DIST = 2f;
-        const float MAX_SQR_DIST = MAX_DIST * MAX_DIST;
+        // const float MAX_DIST = 2f;
+        // const float MAX_SQR_DIST = MAX_DIST * MAX_DIST;
 
         float now = BraveTime.ScaledTimeSinceStartup;
         if (this._nextWiltCheck > now)
@@ -355,15 +374,16 @@ public class FlakseedFlower : MonoBehaviour
         if (!this._sprite)
             return this._wiltRate;
 
-        Vector2 pos = this._body.UnitBottomCenter;
-        foreach (FlakseedFlower f in _ExtantFlowers)
-        {
-            if (!f || f == this || !f._sprite)
-                continue;
-            float sqrMag = (pos - f._sprite.WorldBottomCenter).sqrMagnitude;
-            if (sqrMag < MAX_SQR_DIST)
-                this._wiltRate += (1f - (sqrMag / MAX_SQR_DIST));
-        }
+        //NOTE: wilting nearby flowers is an unnecessary, confusing, and unfun mechanic, so we're just disabling it
+        // Vector2 pos = this._body.UnitBottomCenter;
+        // foreach (FlakseedFlower f in _ExtantFlowers)
+        // {
+        //     if (!f || f == this || !f._sprite)
+        //         continue;
+        //     float sqrMag = (pos - f._sprite.WorldBottomCenter).sqrMagnitude;
+        //     if (sqrMag < MAX_SQR_DIST)
+        //         this._wiltRate += (1f - (sqrMag / MAX_SQR_DIST));
+        // }
 
         #if DEBUG
             // base.gameObject.DrawDebugCircle(pos, MAX_DIST, Color.green.WithAlpha(0.15f));
@@ -405,6 +425,7 @@ public class FlakseedFlower : MonoBehaviour
             return;
         if (!this._grown)
         {
+            DoTrampleVFX();
             UnityEngine.Object.Destroy(base.gameObject);
             return;
         }